requirements/base.txt
django-filter
markdown
djangorestframework
psycopg2
social-auth-app-django
whitenoise
#-e git+https://github.com/rich-hart/django-address.git#egg=django-address
#django-address
